requirements.txt
streamlit==1.37.1
streamlit_navigation_bar
streamlit_lottie
pandas
numpy
scikit-learn==1.4.0
google-generativeai
plotly
supabase
python-dotenv
shap
xgboost
requests
typing
streamlit_shap
matplotlib
pyarrow
umap
imblearn
openpyxl
pygwalker